requirements.txt
python-dotenv == 0.13.0
discord.py == 1.3.3
requests == 2.23.0
pickleDB == 0.9.2